==> notebooks/03_Concatenate_VisiumST_samples.ipynb <==
# Spatial transcriptomics unveils the in situ cellular and molecular hallmarks of the lung in fatal COVID-19

# Concatenate preprocessed Visium ST samples into a single annData object and re-normalise the data.

**Author:** Carlos A. Garcia-Prieto

* This notebook explains the concatenation of the individually preprocessed Visium ST samples into a single object. Because of the concatenation we also need to re-normalise the data.
* See the [concatenation tutorial](https://anndata.readthedocs.io/en/stable/concatenation.html) for more details.
* We have followed single-cell best practices [tutorial](https://www.sc-best-practices.org/preprocessing_visualization/quality_control.html) for preprocessing the concatenated object.

## Import modules

import anndata
import matplotlib.pyplot as plt
import scanpy as sc
import pandas as pd
import seaborn as sns
import scanpy.external as sce
import numpy as np
import os
import anndata2ri
import logging
import seaborn as sns
import rpy2.rinterface_lib.callbacks as rcb
import rpy2.robjects as ro

sc.settings.verbosity = 0
#sc.settings.set_figure_params(
#    dpi=80,
#    facecolor="white",
#    frameon=False,
#)

rcb.logger.setLevel(logging.ERROR)
ro.pandas2ri.activate()
anndata2ri.activate()

%reload_ext rpy2.ipython

pd.set_option('display.max_columns', 500)

## Read Visium ST preprocessed data

# Prepare anndata concat object (read preprocessed objects)
adata_L2P = anndata.read(f'L2P/L2P_preprocessed.h5ad')
adata_L5P = anndata.read(f'L5P/L5P_preprocessed.h5ad')
adata_L14P = anndata.read(f'L14P/L14P_preprocessed.h5ad')
adata_L19P = anndata.read(f'L19P/L19P_preprocessed.h5ad')
adata_L24P = anndata.read(f'L24P/L24P_preprocessed.h5ad')
adata_L3C = anndata.read(f'L3C/L3C_preprocessed.h5ad')
adata_L14C = anndata.read(f'L14C/L14C_preprocessed.h5ad')
adata_L2C = anndata.read(f'L2C/L2C_preprocessed.h5ad')
adata_L11P = anndata.read(f'L11P/L11P_preprocessed.h5ad')
adata_L12P = anndata.read(f'L12P/L12P_preprocessed.h5ad')
adata_CONTROL2 = anndata.read(f'CONTROL2/CONTROL2_preprocessed.h5ad')
adata_HRC2 = anndata.read(f'HRC2/HRC2_preprocessed.h5ad')
adata_HRC4 = anndata.read(f'HRC4/HRC4_preprocessed.h5ad')
adata_HRC5 = anndata.read(f'HRC5/HRC5_preprocessed.h5ad')
adata_HRC6 = anndata.read(f'HRC6/HRC6_preprocessed.h5ad')
adata_HRC8 = anndata.read(f'HRC8/HRC8_preprocessed.h5ad')
adata_HRC10 = anndata.read(f'HRC10/HRC10_preprocessed.h5ad')
adata_HRC11 = anndata.read(f'HRC11/HRC11_preprocessed.h5ad')
adata_HRC12 = anndata.read(f'HRC12/HRC12_preprocessed.h5ad')
adata_HRC13 = anndata.read(f'HRC13/HRC13_preprocessed.h5ad')
adata_HRC16 = anndata.read(f'HRC16/HRC16_preprocessed.h5ad')
adata_HRC17 = anndata.read(f'HRC17/HRC17_preprocessed.h5ad')
adata_HRC18 = anndata.read(f'HRC18/HRC18_preprocessed.h5ad')

#Set metadata
proliferative = ["L2P","L19P","L11P","HRC5","HRC6","HRC8","HRC10","HRC11","HRC12","HRC13","HRC16","HRC17"]
acute = ["L5P","L14P","L24P","L12P","HRC2","HRC4","HRC18"]
controls = ["L3C","L14C","L2C","CONTROL2"]
covid = ["L5P","L14P","L24P","L12P","HRC2","HRC4","HRC18","L2P","L19P","L11P","HRC5","HRC6","HRC8","HRC10","HRC11","HRC12","HRC13","HRC16","HRC17"]
sweden = ["L2P","L19P","L11P","L5P","L14P","L24P","L12P","L3C","L14C","L2C"]
spain = ["CONTROL2","HRC2","HRC4","HRC5","HRC6","HRC8","HRC10","HRC11","HRC12","HRC13","HRC16","HRC17","HRC18"]
batch1 = ["L2P","L5P","L14P","L19P","L24P","L3C","L14C"]
batch2 = ["L2C","L11P","L12P"]
batch3 = ["CONTROL2","HRC2","HRC4","HRC5","HRC6","HRC8","HRC10","HRC11","HRC12","HRC13","HRC16","HRC17","HRC18"]

## Concatenate preprocessed objects

#Concatenate preprocessed objects
adata_concat_hrc_sq = anndata.concat(
    {"L2P" : adata_L2P, "L5P" : adata_L5P, "L14P" : adata_L14P, "L19P" : adata_L19P, "L24P" : adata_L24P, "L3C" : adata_L3C, "L14C" : adata_L14C,
    "L2C" : adata_L2C, "L11P" : adata_L11P, "L12P" : adata_L12P,
    "CONTROL2": adata_CONTROL2, "HRC2": adata_HRC2, "HRC4": adata_HRC4, "HRC5": adata_HRC5,
    "HRC6": adata_HRC6, "HRC8": adata_HRC8, "HRC10": adata_HRC10, "HRC11": adata_HRC11, "HRC12": adata_HRC12, "HRC13": adata_HRC13,
    "HRC16": adata_HRC16, "HRC17": adata_HRC17, "HRC18": adata_HRC18},
    label = "sample",
    index_unique="_",
    join = "outer",
    merge = "unique",
    uns_merge = "unique",
    pairwise = False,
    axis=0,
)

#Explore concatenated object
adata_concat_hrc_sq 

adata_concat_hrc_sq.obs

adata_concat_hrc_sq.var

#adata_concat_hrc_sq.var = adata_concat_hrc_sq.var[["SYMBOL","feature_types","genome"]]

#Add metadata columns
adata_concat_hrc_sq.obs['condition'] = adata_concat_hrc_sq.obs['sample'].isin(covid)
adata_concat_hrc_sq.obs['condition_acute'] = adata_concat_hrc_sq.obs['sample'].isin(acute)
adata_concat_hrc_sq.obs['condition_proliferative'] = adata_concat_hrc_sq.obs['sample'].isin(proliferative)
adata_concat_hrc_sq.obs['origin'] = adata_concat_hrc_sq.obs['sample'].isin(spain)
adata_concat_hrc_sq.obs['batch1'] = adata_concat_hrc_sq.obs['sample'].isin(batch1)
adata_concat_hrc_sq.obs['batch2'] = adata_concat_hrc_sq.obs['sample'].isin(batch2)
adata_concat_hrc_sq.obs['batch3'] = adata_concat_hrc_sq.obs['sample'].isin(batch3)
adata_concat_hrc_sq.obs['batch'] = np.where(adata_concat_hrc_sq.obs['batch1'] == True, 'batch1', (np.where(adata_concat_hrc_sq.obs['batch2'] == True, 'batch2', 'batch3')))
adata_concat_hrc_sq.obs['origin'] = np.where(adata_concat_hrc_sq.obs['origin'] == True, 'Spain', 'Sweden')
adata_concat_hrc_sq.obs['condition'] = np.where(adata_concat_hrc_sq.obs['condition'] == True, 'COVID-19', 'Control')
adata_concat_hrc_sq.obs['condition_subtype'] = np.where(adata_concat_hrc_sq.obs['condition_acute'] == True, 'Acute', (np.where(adata_concat_hrc_sq.obs['condition_proliferative'] == True, 'Proliferative', 'Control')))


adata_concat_hrc_sq.obs[['condition_subtype',"sample"]].value_counts()

### We make a copy of tha anndata object to start working

adata = adata_concat_hrc_sq.copy()

### Because of the concatenation we need to re-normalise the data. Here we just normalise using global scaling by the total counts per cell and shifted logarithm.

adata.layers

adata.X = adata.layers["raw_counts"].copy()

scales_counts = sc.pp.normalize_total(adata, target_sum=None, inplace=False)
# log1p transform
adata.layers["log1p_norm"] = sc.pp.log1p(scales_counts["X"], copy=True)

adata

# Preprocess concatenated data
#### It is always recommended to look at the raw data before performing any integration. This can give some indication of how big any batch effects are and what might be causing them (and therefore which variables to consider as the batch label).

## FEATURE SELECTION USING DEVIANCE

### Deviance works on raw counts so there is no need to replace adata.X with one of the normalized layers.

#TIP when transforming from anndata to R (https://github.com/theislab/anndata2ri/issues/50): 
#Ensure all of the dtypes on any pandas dataframe within your anndata object are NOT object type. 
#Cast any numpy.ndarray objects within your anndata object to scipy.sparse.csc_matrix. 

#Create new object
adata_deviance = adata.copy()

adata_deviance.obsm["spatial"]

#Transform object to int for R import
adata_deviance.obsm["spatial"] = adata_deviance.obsm["spatial"].astype(str).astype(int)

adata_deviance.obsm["spatial"]

# Delete uns as this can contain arbitrary objects which are difficult to convert
del adata_deviance.uns

adata_deviance

%R library(scry)

#Similar to before, we save the AnnData object in our R environment (you may need to run this cell twice, i.e: sometimes it raises an error)
ro.globalenv["adata_deviance"] = adata_deviance

#We can now directly call feature selection with deviance on the non-normalized counts matrix and export the bionomial deviance values as a vector.
%R sce = devianceFeatureSelection(adata_deviance, assay="X")

binomial_deviance = ro.r("rowData(sce)$binomial_deviance").T

#As a next step, we now sort the vector an select the top 6,000 highly deviant genes and save them as an additional column in .var as ‘highly_deviant’. 
#We additionally save the computed binomial deviance in case we want to sub-select a different number of highly variable genes afterwards.

idx = binomial_deviance.argsort()[-6000:]
mask = np.zeros(adata.var_names.shape, dtype=bool)
mask[idx] = True

adata.var["highly_deviant"] = mask
adata.var["binomial_deviance"] = binomial_deviance

#Last, we visualise the feature selection results. We use a scanpy function to compute the mean and dispersion for each gene accross all cells.
sc.pp.highly_variable_genes(adata, layer="log1p_norm")

adata

## Create directory to save results with the concatenated object

folder = "Concatenation_6000_hdg" #We create a folder to save outputs with the concatenated annData

directory = f"{folder}"
# Check if the directory exists
if not os.path.exists(directory):
    # If it doesn't exist, create it
    os.makedirs(directory)

#We inspect our results by plotting dispersion versus mean for the genes and color by ‘highly_deviant’.
with plt.rc_context():  # Use this to set figure params like size and dpi
    sns.scatterplot(data=adata.var, x="means", y="dispersions", hue="highly_deviant", s=5)
    #ax.set_xlim(None, 1.5)
    #ax.set_ylim(None, 3)
    plt.savefig(f"{folder}/{folder}_highly_deviant_genes.png",dpi=300, format="png")
    plt.close()

#We observe that genes with a high mean expression are selected as highly deviant. 

## We will create an object with just the highly deviant (most informative) genes to use for integration.

adata_hdg = adata[:, adata.var["highly_deviant"]].copy()
adata_hdg

### DIMENSIONALITY REDUCTION

#We will use a normalized representation of the dataset for dimensionality reduction and visualization, specifically the shifted logarithm.
adata_hdg.X = adata_hdg.layers["log1p_norm"]

### PCA

# setting highly variable as highly deviant
#adata_hdg.var["highly_variable"] = adata_hdg.var["highly_deviant"]

adata_hdg.var[["highly_variable", "highly_deviant"]].value_counts()

sc.pp.pca(adata_hdg, svd_solver="arpack")

### t-SNE

sc.tl.tsne(adata_hdg, use_rep="X_pca")

### UMAP

#We first calculate PCA and subsequently a neighborhood graph on our data.
sc.pp.neighbors(adata_hdg)
sc.tl.umap(adata_hdg)

### Inspecting potential batch effect to identify batch variables to consider for integration

#We can now also inspect the batch effect in our PCA, TSNE or UMAP plot and potentially identify batch variables.
with plt.rc_context():  
    plt.rcParams["figure.figsize"] = (4, 4)
    sc.pl.pca_scatter(adata_hdg,color=["batch", "condition_subtype", "sample"], show=False, wspace=1)
    plt.savefig(f"{folder}/{folder}_PCA_dimensionality_reduction_hdg_batch.png",dpi=300, format="png")
    plt.close()

with plt.rc_context():  
    plt.rcParams["figure.figsize"] = (4, 4)
    sc.pl.tsne(adata_hdg,color=["batch", "condition_subtype", "sample"], show=False, wspace=1)
    plt.savefig(f"{folder}/{folder}_tSNE_dimensionality_reduction_hdg_batch.png",dpi=300, format="png")
    plt.close()

with plt.rc_context():  
    plt.rcParams["figure.figsize"] = (4, 4)
    sc.pl.umap(adata_hdg,color=["batch", "condition_subtype", "sample"], show=False, wspace=1)
    plt.savefig(f"{folder}/{folder}_UMAP_dimensionality_reduction_hdg_batch.png",dpi=300, format="png")
    plt.close()

### CLUSTERING

#We call Leiden algorithm
sc.tl.leiden(adata_hdg, key_added="leiden_concat_hdg_res1", resolution=1.0)

#The default resolution parameter in scanpy is 1.0. 
#However, in many cases the analyst may want to try different resolution parameters to control the coarseness of the clustering. 
sc.tl.leiden(adata_hdg, key_added="leiden_concat_hdg_res0_25", resolution=0.25)
sc.tl.leiden(adata_hdg, key_added="leiden_concat_hdg_res0_5", resolution=0.5)
#sc.tl.leiden(adata, key_added="leiden_res1", resolution=1.0)

#Visualize clustering
with plt.rc_context():  # Use this to set figure params like size and dpi
    plt.rcParams["figure.figsize"] = (4, 4)
    sc.pl.umap(adata_hdg,color=["leiden_concat_hdg_res0_25", "leiden_concat_hdg_res0_5", "leiden_concat_hdg_res1"],legend_loc="on data", show=False)
    plt.savefig(f"{folder}/{folder}_leiden_community_res_hdg.png",dpi=300, format="png")
    plt.close()

### Save concatenated anndata object with highly deviant genes

adata_hdg.write_h5ad(f'{folder}/Concatenation_adata_hdg_6000.h5ad', compression='gzip')
