==> notebooks/04_HLCA_Core_Processing.ipynb <==
# Spatial transcriptomics unveils the in situ cellular and molecular hallmarks of the lung in fatal COVID-19

# Process Human Lung Cell Atlas (HLCA) core object to use it as reference dataset for deconvolution and mapping of Visium ST data. 

**Author:** Carlos A. Garcia-Prieto

* This notebook explains the processing of the HLCA core annData object to use it as a reference single-cell dataset for deconvolution and mapping of the studied Visium ST samples.
* See the [HLCA paper](https://doi.org/10.1038/s41591-023-02327-2) for more details and to download [HLCA core annData object.](https://cellxgene.cziscience.com/collections/6f6d381a-7701-4781-935c-db10d30de293)


## Import modules

import anndata
import matplotlib.pyplot as plt
import scanpy as sc
import pandas as pd
import seaborn as sns
import scanpy.external as sce
import numpy as np

pd.set_option('display.max_rows', 250)
pd.set_option('display.max_columns', 500)

#Read HLCA core annData object
HLCA_core = anndata.read(f'/Users/carlosgarciaprieto/Proyectos_IJC/Spatial/COVID/Results/Python/HLCA_publication/HLCA_core.h5ad')

#Explore HLCA core
HLCA_core

HLCA_core.obs

HLCA_core.obs.columns

HLCA_core.obs["assay"].value_counts()

HLCA_core.obs["sequencing_platform"].value_counts()

HLCA_core.obs["donor_id"].value_counts()

HLCA_core.obs["tissue"].value_counts()

HLCA_core.obs[["tissue","anatomical_region_ccf_score"]].value_counts()

## We select only lung parenchyma cells to match our studied Visium ST samples

HLCA_core_parenchyma = HLCA_core[HLCA_core.obs["tissue"] == "lung parenchyma"].copy()

#Explora HLCA parenchyma object
HLCA_core_parenchyma

#Explore finest annotation level
HLCA_core_parenchyma.obs["ann_finest_level"].value_counts()

## We select only cell types with at least 150 cells (finest annotation level) present in the HLCA core lung parenchyma object for a more robust reference model training

min_cells = 150

counts = HLCA_core_parenchyma.obs.groupby('ann_finest_level')['ann_finest_level'].count()

minimumValueCells = counts[counts>=min_cells]

HLCA_core_parenchyma_filter = HLCA_core_parenchyma[HLCA_core_parenchyma.obs['ann_finest_level'].isin(minimumValueCells.index)]

HLCA_core_parenchyma_filter

HLCA_core_parenchyma_filter.obs["ann_finest_level"].value_counts()

## Save HLCA filtered object for reference model training

#Write filtered lung parenchyma annData object
HLCA_core_parenchyma_filter.write_h5ad('HLCA_publication/HLCA_core_parenchyma_filter.h5ad', compression='gzip')
