==> notebooks/16_c2l_cell_type_specific_gene_expr_spatial.ipynb <==
# Spatial transcriptomics unveils the in situ cellular and molecular hallmarks of the lung in fatal COVID-19

# Pneumocytes specific expression of CAV1 in spatial coordinates

**Author:** Carlos A. Garcia-Prieto

* This notebook explains how to plot cell type specific expression of genes in spatial coordinates using cell2location 
* We followed cell2location plotting cell-type specific expression of genes in spatial coordinates [tutorial.](https://cell2location.readthedocs.io/en/latest/notebooks/cell2location_tutorial.html#Estimate-cell-type-specific-expression-of-every-gene-in-the-spatial-data-(needed-for-NCEM)) 

## Import modules

import warnings
warnings.filterwarnings("ignore")
import matplotlib.pyplot as plt
import scanpy as sc
import numpy as np
import os
import cell2location
from cell2location.utils import select_slide
import matplotlib as mpl

#Set data directory
results_folder = "/mnt/beegfs/cgarcia/Spatial/COVID19/cell2location/HLCA_publication/HLCA/"
ref_run_name = f'{results_folder}reference_signatures_finest'
run_name = f'{results_folder}cell2location_map_finest'

## Read cell type specific gene expression

#Read cell type specific gene expression needed for NCEM
adata_file = f"{run_name}/ad_vis_post_distrib_finest_cell_type_gene_expr.h5ad"
adata_vis = sc.read_h5ad(adata_file)

# Plot cell type specific expression of CAV1 gene in spatial coordinates
## Below we plot the cell type specific expression of CAV1 compared to total expression of the gene (first panel).


Note that `plot_genes_per_cell_type` function often need customization so it is not included into cell2location package - you need to copy it from https://github.com/BayraktarLab/cell2location/blob/master/docs/notebooks/tutorial_utils.py to use on your system.

def plot_genes_per_cell_type(slide, genes, ctypes):

    n_genes = len(genes)
    n_ctypes = len(ctypes)
    fig, axs = plt.subplots(
        nrows=n_genes, ncols=n_ctypes + 1, figsize=(4.5 * (n_ctypes + 1) + 2, 5 * n_genes + 1), squeeze=False
    )
    # axs = axs.reshape((n_genes, n_ctypes+1))

    # plots of every gene
    for j in range(n_genes):

        # limit color scale at 99.9% quantile of gene expression (computed across cell types)
        quantile_across_ct = np.array(
            [
                np.quantile(slide.layers[n][:, slide.var["SYMBOL"] == genes[j]].toarray(), 0.999) 
                for n in slide.uns["mod"]["factor_names"]
            ]
        )
        quantile_across_ct = np.partition(quantile_across_ct.flatten(), -2)[-2]
        sc.pl.spatial(
            slide,
            cmap="magma",
            color=genes[j],
            # layer=ctypes[i],
            gene_symbols="SYMBOL",
            ncols=4,
            size=1.3,
            img_key="hires",
            # limit color scale at 99.9% quantile of gene expression
            vmin=0,
            vmax="p99.9",
            ax=axs[j, 0],
            show=False,
        )

        # plots of every cell type
        for i in range(n_ctypes):
            sc.pl.spatial(
                slide,
                cmap="magma",
                color=genes[j],
                layer=ctypes[i],
                gene_symbols="SYMBOL",
                ncols=4,
                size=1.3,
                img_key="hires",
                # limit color scale at 99.9% quantile of gene expression
                vmin=0,
                vmax=quantile_across_ct,
                ax=axs[j, i + 1],
                show=False,
            )
            axs[j, i + 1].set_title(f"{genes[j]} {ctypes[i]}")

    return fig, axs

## Plot cell type specific gene expression

#Create directory for Plot
directory = f'{run_name}/Plots'
# Check if the directory exists
if not os.path.exists(directory):
    # If it doesn't exist, create it
    os.makedirs(directory)

#Select gene and cell types
gene = ['CAV1'] 
ctypes = ['AT1', 'AT2', 'AT0']

#Set sample used for spatio temporal trajectory analysis
sample = "HRC11"

#Select sample
slide = select_slide(adata_vis, f'{sample}')

<div class="alert alert-info">
<b>Paper Figure!</b>
Panel Figure 5C
</div>

#Plot CAV1 cell type specific expression
with mpl.rc_context({'axes.facecolor':  'white', 'figure.figsize': [5, 5]}):
    #from tutorial_utils import plot_genes_per_cell_type
    plot_genes_per_cell_type(slide, gene, ctypes)
    plt.savefig(f"{run_name}/Plots/{sample}_pneumocytes_{gene}_gene_expression_Paper.png",dpi=300, format="png",pad_inches=0.2,bbox_inches="tight")
    plt.close()

### Modules and their versions used for this analysis


Useful for debugging and reporting issues.

cell2location.utils.list_imported_modules()
